implementado, testes pendentes

+++ src/DServer.cpp
@@ -1,14 +1,13 @@
 #include "../include/DServer.h"
 
-DServer::DServer()
+DServer::DServer(string confFile)
 {
 	serverSocket = new DSocket();
 	if(serverSocket->isOpen()) {
 		bool bindOk = serverSocket->bindSocket("localhost",SERVER_PORT);
 		if(bindOk) {
 			for(int i = 0; i < MAX_CONNECTIONS; i++) ports[i] = AVAILABLE;
-			this->initialize();
-			isWorking = true; }
+			isWorking = this->initialize() && this->loadConfFile(); && this->initializeRing() }
 		else {
 			cout << "DServer::DServer() - Erro. Chamada de serverSocket->bind() retornou código de erro." << endl;
 			serverSocket->closeSocket();
@@ -18,7 +17,7 @@ DServer::DServer()
 		isWorking = false; }
 }
 
-void DServer::initialize()
+bool DServer::initialize()
 {
 	this->homeDir = string("/tmp");
 	string serverPath = homeDir + "/dbox-server";
@@ -35,8 +34,101 @@ void DServer::initialize()
 				string clientName(de->d_name);
 				DClient* newClient = new DClient(clientName);
 				clients.push_back(newClient);
-				newClient->fillFilesList("/dbox-server/" + clientName); } } }
-	else if(mkdir(serverPath.c_str(), 0777) != 0) cout << "DServer::initialize() - Erro. Não conseguiu criar diretório raiz do servidor." << endl;
+				newClient->fillFilesList("/dbox-server/" + clientName); } } 
+			return true; }
+	else if(mkdir(serverPath.c_str(), 0777) != 0) {
+		cout << "DServer::initialize() - Erro. Não conseguiu criar diretório raiz do servidor." << endl;
+		return false; }
+	else
+		return true;
+}
+
+bool DServer::loadConfFile(string confFile)
+{
+	cout << "DServer::loadConfFile() - Lendo endereços de \"" << confFile << "\"." << endl;
+	thisServerAddresses = getIpAddressList();
+	ifstream infile;
+	try {
+		infile.open(confFile); }
+	catch (const exception &e) {
+		cout << "DServer::loadConfFile() - Exception. Lançado " << e.what() << " ao abrir o arquivo de configuração para leitura." << endl;
+		return false; }
+
+	string confLine;
+	int numServers = 0;
+	string ip;
+	int port;
+	thisServerID = -1;
+	while (!infile.eof()) {
+		infile >> confLine;
+		if (confLine != "") {
+			int colon = confLine.find(':');
+			try {
+				ip = confLine.substr(0, colon);
+				port = stoi(confLine.substr(colon+1)); }
+			catch (const exception &e) {
+				cout << "DServer::loadConfFile() - Exception. Lançado " << e.what() << " processando endereços de servidor." << endl;
+				return false; }
+			if (!validateAddress(ip, port))
+				cout << "DServer::loadConfFile() - Erro. \"" << confLine << "\" não é um endereço ip:porta válido." << endl;
+			if (thisServerAddresses.find(ip) != string::npos)
+				thisServerID = numServers;
+			pair<string, int> newServer = make_pair(ip, port);
+			servers.push_back(newServer);
+
+			numServers++;
+			cout << "DServer::loadConfFile() - Servidor #" << numServers << "  " << ip << ":" << port;
+			if (thisServerID == numServers - 1)
+				cout << " (este servidor)";
+			cout << endl; } }
+	if (thisServerID == -1) {
+		cout << "DServer::loadConfFile() - Erro. Arquivo de configuração não tem entrada para este servidor." << endl; 
+		return false; }
+	if (numServers == 0) {
+		cout << "DServer::loadConfFile() - Erro. Nenhum servidor adicionado." << endl; 
+		return false; }
+	else if (numServers == 1)
+		cout << "DServer::loadConfFile() - Aviso. Apenas um servidor encontrado. Recuperação em caso de falhas não será possível." << endl;
+
+	serverStatus = OFFLINE;
+	return true;
+}
+
+string DServer::getIpAddressList() {
+	struct ifaddrs *ifap, *ifa;
+	struct sockaddr_in *sa;
+	char *addr;
+	string r = "|";
+
+	getifaddrs (&ifap);
+	for (ifa = ifap; ifa; ifa = ifa->ifa_next) {
+		if (ifa->ifa_addr->sa_family==AF_INET) {
+			sa = (struct sockaddr_in *) ifa->ifa_addr;
+			addr = inet_ntoa(sa->sin_addr);
+			if (addr != "127.0.0.1")
+				r += addr + "|"; } }
+
+	freeifaddrs(ifap);
+	return r;
+}
+
+bool DServer::initializeRing()
+{
+	upRingSocket = new DSocket(TCP);
+	if(!upRingSocket->bindSocket("localhost",SERVER_PORT-1)) {
+		cout << "DServer::setUpRing() - Erro. Chamada de upRingSocket->bind() retornou código de erro." << endl;
+		upRingSocket->closeSocket();
+		return false; }
+	if(!upRingSocket->listen()) {
+		cout << "DServer::setUpRing() - Erro. Chamada de upRingSocket->listen() retornou código de erro." << endl;
+		upRingSocket->closeSocket();
+		return false; }
+	upRingConnectedSocket = NULL;
+	DownRingSocket = new DSocket(TCP);
+	isUpRingConnected = false;
+	isDownRingConnected = false;
+	filePartsRemaining = 0;
+	return true:
 }
 
 DClient* DServer::findClient(string clientName)
@@ -46,6 +138,126 @@ DClient* DServer::findClient(string clientName)
 		DClient* client = *(it);
 		if(client->getName() == clientName) return client; }
 	return NULL;
+}
+
+void DServer::coordinate()
+{
+	while(true) {
+		if (!isRingConnected)
+			if (!connectRing())
+				cout << "DServer::coordinate() - Erro. Não foi possível estabelecer conexão em anel." << endl;
+		if (serverStatus == MASTER) {
+			if (ringMessage) {
+				DMessage* validationMessage;
+				if (upRingConnectedSocket->receive(&validationMessage)) {
+					if (validationMessage == ringMessage)
+						delete RingMessage;
+						ringMessage = NULL;
+					else if ((errno != EAGAIN) && (errno != EWOULDBLOCK)) {
+						cout << "DServer::coordinate() - Erro. Erro ao checar a integridade da mensagem." << endl;
+						isUpRingConnected = false; }
+					delete validationMessage; } }
+			if (!rinqQueue.empty()) {
+				if (downRingSocket->send(ringQueue.top())) {
+					delete ringQueue.top();
+					ringQueue.pop(); }
+				else if ((errno != EAGAIN) && (errno != EWOULDBLOCK)) {
+					cout << "DServer::coordinate() - Erro. downRingSocket->send() retornou código de erro. Envio será tentado novamente se a conexão estiver aberta." << endl;
+					isDownRingConnected = false; } } }
+		else if (serverStatus == SLAVE) {
+			if (!ringMessage)
+				if (upRingConnectedSocket->receive(&ringMessage))
+					if (!processUpdate()) {
+						cout << "DServer::coordinate() - Erro. Não foi possível processar a atualização recebida." << endl;
+						breakRing(); }
+				else if ((errno != EAGAIN) && (errno != EWOULDBLOCK)) {
+						cout << "DServer::coordinate() - Erro. upRingSocket->receive() retornou código de erro. Recebimento será tentado novamente se a conexão estiver aberta." << endl;
+						isUpRingConnected = false; }
+			if (ringMessage)
+				if (downRingSocket->send(ringMessage))
+						delete ringMessage;
+						ringMessage = NULL;
+				else if ((errno != EAGAIN) && (errno != EWOULDBLOCK)) {
+					cout << "DServer::coordinate() - Erro. Chamada de downRingSocket->send() retornou código de erro. Envio será tentado novamente se a conexão estiver aberta." << endl;
+					isDownRingConnected = false; } }
+		else if (serverStatus == OFFLINE) {
+			cout << "DServer::coordinate() - Servidor está offline." << endl;
+			breakRing(); } } }
+
+bool DServer::connectRing()
+{
+	if (!isUpRingConnected) {
+		if (upRingConnectedSocket)
+			delete upRingConnectedSocket;
+		if (upRingConnectedSocket = upRingSocket->acceptConnection())
+			isUpRingConnected = true;
+		if (getSocketServerID(upRingConnectedSocket) > thisServerID)
+			serverStatus = MASTER;
+		else
+			serverStatus = SLAVE; }
+
+	if (!isDownRingConnected) {
+		int i = thisServerID + 1;
+		while ((!isDownRingConnected) && (i != thisServerID)) {
+			for (int j = 0; j < MAX_RETRIES; j++) {
+				cout << "DServer::connectRing() - Conectando ao servidor #" << (i + 1) << " - "
+					 << get<POS_IP>(servers[i])  << ":"
+					 << get<POS_PORT>(servers[i])  << endl;
+				downRingSocket->setDestination(get<POS_IP>servers[i], get<POS_PORT>(servers[i]));
+				if (downRingSocket->connectSocket())
+					isDownRingConnected = true; }
+			id += 1 % servers.size(); }
+		if (!isDownRingConnected) {
+			cout << "DServer::connectRing() - Nenhum servidor aceitou conexão." << endl;
+			serverStatus = OFFLINE; } }
+}
+
+int DServer::getSocketServerID(DSocket* socket)
+{
+	string thisIp = socket.getDestinationIp;
+	int thisPort = socket.getDestinationPort;
+	for (int i = 0; i < servers.size(); i++)
+		if (thisIp == get<POS_IP>(servers[i])) && (thisPort == get<POS_PORT>(servers[i]))
+			return i;
+	return -1;
+}
+
+bool DServer::processUpdate()
+{
+	ofstream receivingFile;
+	if (filePartsRemaining > 0) {
+		receivingFile.open(receivingFilePath, ofstream::app | ofstream::binary);
+		receivingFile.write(ringMessage->get(),ringMessage->length());
+		receivingFile.close();
+		filePartsRemaining--; }
+	else if (ringMessage.toString().substr(0,12) == "receive file") {
+		receivingFilePath = ringMessage.toString;
+		receivingFilePath = receivingFilePath.substr(receivingFilePath.find(" ")+1);
+		string fileParts = fileName;
+		receivingFilePath = receivingFilePath.substr(receivingFilePath.find(" ")+1);
+		fileParts = fileParts.substr(0, fileParts.find(" "));
+		receivingFilePartsRemaining = stoi(fileParts);
+		receivingFile.open(receivingFilePath, ofstream::trunc | ofstream::binary);
+		receivingFile.close(); }
+	else if (ringMessage.toString().substr(0,11) == "delete file") {
+		string filePath = ringMessage.toString;
+		filePath = filePath.substr(fileName.find(" ")+1);
+		remove(filePath.c_str()) }
+	else {
+		cout << "DServer::processUpdate() - Erro. Comando \"" << ringMessage.toString.substr(0,20);
+		if (ringMessage.toString.size() > 20)
+			cout << "...";
+		cout << "\" não identificado." << endl; }
+}
+
+void DServer::breakRing()
+{
+	if (upRingSocket)
+		upRingSocket->close();
+	if (downRingSocket)
+		downRingSocket->close();
+	if (downRingConnectedSocket)
+		downRingConnectedSocket->close();
 }
 
 void DServer::listen()
@@ -81,7 +293,9 @@ void DServer::listen()
 					cout << "Conexão recusada. Cliente já está conectado neste dispositivo." << endl;
 					serverSocket->reply(new DMessage("connection refused"));
 					continue; }
-				else acceptConnection(client, serverSocket->getSenderIp(), serverSocket->getSenderPort()); } } }
+				else acceptConnection(client, serverSocket->getSenderIp(), serverSocket->getSenderPort()); } }
+			else if (connectionRequest->toString() == "ping")
+				serverSocket->reply(new DMessage("ok")); }
 }
 
 void DServer::acceptConnection(DClient* client, struct in_addr clientIp, unsigned short clientPort)
@@ -106,7 +320,7 @@ void DServer::acceptConnection(DClient* client, struct in_addr clientIp, unsigne
 			client->getConnectionsList()->push_back(newConnectionSocket);
 			thread* newConnection = new thread(&DServer::messageProcessing, this, client, newConnectionSocket);
 			connections.push_back(newConnection);
-			bool ackSent = serverSocket->reply(new DMessage("ack " + to_string(portNumber)));
+			bool ackSent = serverSocket->reply(new DMessage("ack port" + to_string(portNumber)));
 			if(ackSent) {
 				cout << "Conexão aceita. Cliente: " << client->getName() << ". IP: " << inet_ntoa(clientIp) << ". Porta: " << ntohs(clientPort) << "." << endl;
 				break; }
@@ -167,7 +381,7 @@ void DServer::sendFile(DClient* client, DSocket* connection, DMessage* message)
 		file.seekg(filePointer, file.beg);
 		filePointer += BUFFER_SIZE;
 		if(packetsSent == totalPackets) blockSize = lastPacketSize;
-		file.read(packetContent, blockSize);					
+		file.read(packetContent, blockSize);
 		DMessage* packet = new DMessage(packetContent,blockSize);
 		bool packetSent = connection->reply(packet);
 		if(!packetSent) {
@@ -209,6 +423,7 @@ void DServer::deleteFile(DClient* client, DSocket* connection, DMessage* message
 		return; }
 	DMessage* removed = new DMessage("removed");
 	connection->reply(removed);
+	ringQueue.push(new DMessage("delete file " + filePath));
 	client->getFilesList()->clear();
 	client->fillFilesList("/dbox-server/" + client->getName());
 	cout << "Arquivo deletado do servidor: " << fileName << ". Cliente: " << client->getName() << "." << endl;
@@ -233,6 +448,7 @@ void DServer::receiveFile(DClient* client, DSocket* connection, DMessage* messag
 	int packetsReceived = 0;
 	int packetSize = 0;
 	DMessage* packet = NULL;
+	ringQueue.push(new DMessage("receive file " totalPackets + " " filePath));
 	while(packetsReceived < totalPackets) {
 		packetsReceived++;				
 		bool packetReceived = connection->receive(&packet);
@@ -246,7 +462,8 @@ void DServer::receiveFile(DClient* client, DSocket* connection, DMessage* messag
 			cout << "DServer::receiveFile() - Erro ao enviar confirmação de recebimento de pacote de dados. Recebimento interrompido." << endl;
 			newFile.close();
 			return; }
-		newFile.write(packet->get(),packet->length()); }
+		newFile.write(packet->get(),packet->length());
+		ringQueue.push(new DMessage(packet)); }
 	newFile.close();
 	DMessage* lastModificationTime = NULL;
 	bool lmtReceived = connection->receive(&lastModificationTime);
@@ -320,32 +537,31 @@ bool DServer::closeConnection(DClient* client, DSocket* connection)
 
 void DServer::messageProcessing(DClient* client, DSocket* connection)
 {
-	while(true) {
-		DMessage* message = NULL;
-		bool messageReceived = connection->receive(&message);
-		if(messageReceived) {
-			if(message->toString() == "close connection") {
-				mtxClientsListUpdate.lock();
-				bool connectionClosed = closeConnection(client,connection);
-				mtxClientsListUpdate.unlock();
-				if(connectionClosed) break;	}	 
-			if(message->toString().substr(0,7) == "receive") {
-				client->getMutex()->lock();
-				receiveFile(client, connection, message);
-				client->getMutex()->unlock(); }
-			if(message->toString().substr(0,6) == "delete") {
-				client->getMutex()->lock();
-				deleteFile(client, connection, message);
-				client->getMutex()->unlock(); }
-			if(message->toString().substr(0,4) == "send") {
-				client->getMutex()->lock();
-				sendFile(client, connection, message);
-				client->getMutex()->unlock(); }
-			if(message->toString() == "list") {
-				client->getMutex()->lock();
-				listFiles(client, connection, message);
-				client->getMutex()->unlock(); }
-			if (message->toString() == "ping")
-				connection->reply(new DMessage("ok")); }
-		else continue; }
+	while(true)
+		if (serverStatus == MASTER) {
+			DMessage* message = NULL;
+			bool messageReceived = connection->receive(&message);
+			if(messageReceived) {
+				if(message->toString() == "close connection") {
+					mtxClientsListUpdate.lock();
+					bool connectionClosed = closeConnection(client,connection);
+					mtxClientsListUpdate.unlock();
+					if(connectionClosed) break;	}
+				if(message->toString().substr(0,7) == "receive") {
+					client->getMutex()->lock();
+					receiveFile(client, connection, message);
+					client->getMutex()->unlock(); }
+				if(message->toString().substr(0,6) == "delete") {
+					client->getMutex()->lock();
+					deleteFile(client, connection, message);
+					client->getMutex()->unlock(); }
+				if(message->toString().substr(0,4) == "send") {
+					client->getMutex()->lock();
+					sendFile(client, connection, message);
+					client->getMutex()->unlock(); }
+				if(message->toString() == "list") {
+					client->getMutex()->lock();
+					listFiles(client, connection, message);
+					client->getMutex()->unlock(); } }
+			else continue; }
 }
